+-upd: moved messages body+footer+header
Also I created NoTickedFound.tsx
