+++ app/(site)/support/tickets/[ticketId]/components/MessagesFooter.tsx
