+++ app/(site)/support/tickets/[ticketId]/components/NoTicketFound.tsx
@@ -1,0 +1,9 @@
+export function NoTicketFound() {
+  return (
+    <div
+      className="bg-foreground-accent w-full h-full justify-center items-center
+    shadow-[inset_0px_8px_6px_rgba(0,0,0,0.4)] z-[100]">
+      No ticket with this id found
+    </div>
+  )
+}
