+++ app/(site)/support/tickets/[ticketId]/components/index.ts
@@ -1,0 +1,4 @@
+export { MessagesHeader } from "../components/MessagesHeader"
+export { MessagesBody } from "../components/MessagesBody"
+export { MessagesFooter } from "../components/MessagesFooter"
+export { NoTicketFound } from "./NoTicketFound"
